+++ chats/apps/api/v1/rooms/viewsets.py
@@ -485,16 +485,24 @@ class RoomViewset(
             action = "transfer" if self.request.user.email != user else "pick"
             feedback = create_transfer_json(
                 action=action,
-                from_=old_instance.user or old_instance.queue,
+                from_=old_user or old_instance.queue,
                 to=instance.user,
                 requested_by=self.request.user,
             )
+
+            create_room_feedback_message(
+                instance,
+                feedback,
+                method=RoomFeedbackMethods.ROOM_TRANSFER,
+                requested_by=self.request.user,
+            )
+            instance.add_transfer_to_history(feedback)
 
         if queue:
             # Create constraint to make queue not none
             feedback = create_transfer_json(
                 action="transfer",
-                from_=old_instance.user or old_instance.queue,
+                from_=old_user or old_instance.queue,
                 to=instance.queue,
                 requested_by=self.request.user,
             )
@@ -509,17 +517,16 @@ class RoomViewset(
             room_metric.transfer_count += 1
             room_metric.save()
 
-        instance.save()
-        instance.add_transfer_to_history(feedback)
+            create_room_feedback_message(
+                instance,
+                feedback,
+                method=RoomFeedbackMethods.ROOM_TRANSFER,
+                requested_by=self.request.user,
+            )
 
-        # Create a message with the transfer data and Send to the room group
-        # TODO separate create message in a function
-        create_room_feedback_message(
-            instance,
-            feedback,
-            method=RoomFeedbackMethods.ROOM_TRANSFER,
-            requested_by=self.request.user,
-        )
+            instance.add_transfer_to_history(feedback)
+
+        instance.save()
 
         if old_user is None and user:  # queued > agent
             instance.notify_queue("update")