+++ chats/apps/projects/models/models.py
@@ -351,6 +351,16 @@ class ProjectPermission(BaseSoftDeleteModel, BaseModel):
         help_text=_("Last time the agent sent a ping (heartbeat)"),
     )
 
+    custom_rooms_limit = models.PositiveIntegerField(
+        _("Custom rooms limit"),
+        null=True,
+        blank=True,
+    )
+    is_custom_limit_active = models.BooleanField(
+        _("Is custom limit active?"),
+        default=False,
+    )
+
     objects = UserPermissionsManager()
     auth = models.Manager()